Refactor airflow plugin
