+++ openmetadata-airflow-apis/openmetadata_managed_apis/api/routes/__init__.py
@@ -1,0 +1,7 @@
+from os.path import dirname, basename, isfile, join
+import glob
+
+modules = glob.glob(join(dirname(__file__), "*.py"))
+__all__ = [
+    basename(f)[:-3] for f in modules if isfile(f) and not f.endswith("__init__.py")
+]
