+++ openmetadata-airflow-apis/openmetadata_managed_apis/plugins/dag_templates/dag_runner.j2
@@ -1,0 +1,9 @@
+"""
+This file has been generated from dag_runner.j2
+"""
+from airflow import DAG
+from openmetadata_managed_apis.workflows import workflow_factory
+
+workflow = workflow_factory.WorkflowFactory.create("{{ workflow_config_file }}")
+workflow.generate_dag(globals())
+dag = workflow.get_dag()
