+++ openmetadata-airflow-apis/setup.py
@@ -8,10 +8,28 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
-
+import glob
 import os
+import pathlib
+from itertools import chain
 
-from setuptools import find_namespace_packages, setup
+from setuptools import setup, find_packages
+
+PLUGIN_NAME = "openmetadata_managed_apis"
+PLUGIN_ROOT_FILE = "main"
+PLUGIN_ENTRY_POINT = "RestApiPlugin"
+HOME = pathlib.Path(__file__).parent
+README = (HOME / "README.md").read_text()
+CONFIG = HOME / PLUGIN_NAME / "config.py"
+
+
+def get_package_data():
+    extensions = ["js", "css", "html", "j2"]
+
+    files = list(
+        chain(*[glob.glob(f"{PLUGIN_NAME}/**/*.%s" % x, recursive=True) for x in extensions])
+    )
+    return [x.split(os.sep, 1)[1] for x in files]
 
 
 def get_long_description():
@@ -22,17 +40,17 @@ def get_long_description():
 
 
 base_requirements = {
-    "openmetadata-ingestion[airflow-container]~=0.10",
+    "openmetadata-ingestion[airflow-container]~=0.11",
     "pendulum~=2.1.2",
     "packaging~=21.2",
     "setuptools~=58.3.0",
-    "apache-airflow==2.1.4",
-    "Flask~=1.1.4",
+    "apache-airflow==2.3.3",
+    "Flask~=2.1",
     "Flask-Admin==1.6.0",
 }
 
 dev_requirements = {
-    "black==21.12b0",
+    "black==22.3.0",
     "pytest",
     "pylint",
     "pytest-cov",
@@ -41,7 +59,10 @@ dev_requirements = {
 }
 
 setup(
-    name="openmetadata-airflow-managed-apis",
+    name=PLUGIN_NAME,
+    packages=find_packages(include=[f"{PLUGIN_NAME}.*", PLUGIN_NAME]),
+    include_package_data=True,
+    package_data={PLUGIN_NAME: get_package_data()},
     version="0.12.0.dev0",
     url="https://open-metadata.org/",
     author="OpenMetadata Committers",
@@ -49,15 +70,18 @@ setup(
     description="Airflow REST APIs to create and manage DAGS",
     long_description=get_long_description(),
     long_description_content_type="text/markdown",
+    entry_points={
+        "airflow.plugins": [
+            f"{PLUGIN_NAME} = {PLUGIN_NAME}.{PLUGIN_ROOT_FILE}:{PLUGIN_ENTRY_POINT}"
+        ]
+    },
     python_requires=">=3.7",
-    package_dir={"": "src"},
     zip_safe=False,
     dependency_links=[],
     project_urls={
         "Documentation": "https://docs.open-metadata.org/",
         "Source": "https://github.com/open-metadata/OpenMetadata",
     },
-    packages=find_namespace_packages(where="./src", exclude=["tests*"]),
     install_requires=list(base_requirements),
     extras_require={
         "base": list(base_requirements),
