--- openmetadata-airflow-apis/src/openmetadata/helpers.py
@@ -1,24 +1,0 @@
-#  Copyright 2021 Collate
-#  Licensed under the Apache License, Version 2.0 (the "License");
-#  you may not use this file except in compliance with the License.
-#  You may obtain a copy of the License at
-#  http://www.apache.org/licenses/LICENSE-2.0
-#  Unless required by applicable law or agreed to in writing, software
-#  distributed under the License is distributed on an "AS IS" BASIS,
-#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-#  See the License for the specific language governing permissions and
-#  limitations under the License.
-"""
-Helper functions
-"""
-import re
-from typing import Optional
-
-
-def clean_dag_id(raw_dag_id: Optional[str]) -> Optional[str]:
-    """
-    Given a string we want to use as a dag_id, we should
-    give it a cleanup as Airflow does not support anything
-    that is not alphanumeric for the name
-    """
-    return re.sub("[^0-9a-zA-Z-_]+", "_", raw_dag_id) if raw_dag_id else None
