--- openmetadata-airflow-apis/src/openmetadata/workflows/ingestion/registry.py
@@ -1,31 +1,0 @@
-#  Copyright 2021 Collate
-#  Licensed under the Apache License, Version 2.0 (the "License");
-#  you may not use this file except in compliance with the License.
-#  You may obtain a copy of the License at
-#  http://www.apache.org/licenses/LICENSE-2.0
-#  Unless required by applicable law or agreed to in writing, software
-#  distributed under the License is distributed on an "AS IS" BASIS,
-#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-#  See the License for the specific language governing permissions and
-#  limitations under the License.
-"""
-DAG builder registry.
-
-Add a function for each type from PipelineType
-"""
-
-
-from openmetadata.workflows.ingestion.metadata import build_metadata_dag
-from openmetadata.workflows.ingestion.profiler import build_profiler_dag
-from openmetadata.workflows.ingestion.usage import build_usage_dag
-
-from metadata.generated.schema.entity.services.ingestionPipelines.ingestionPipeline import (
-    PipelineType,
-)
-from metadata.utils.dispatch import enum_register
-
-build_registry = enum_register()
-
-build_registry.add(PipelineType.metadata.value)(build_metadata_dag)
-build_registry.add(PipelineType.usage.value)(build_usage_dag)
-build_registry.add(PipelineType.profiler.value)(build_profiler_dag)
