--- openmetadata-airflow-apis/src/plugins/dag_templates/dag_runner.j2
@@ -1,9 +1,0 @@
-"""
-This file has been generated from dag_runner.j2
-"""
-from airflow import DAG
-from openmetadata.workflows import workflow_factory
-
-workflow = workflow_factory.WorkflowFactory.create("{{ workflow_config_file }}")
-workflow.generate_dag(globals())
-dag = workflow.get_dag()
